Split out the command line stuff to own file
